Updates the doc of Swagger

+++ models/book_model.go
@@ -4,7 +4,7 @@ type Book struct {
 	Id          int     `json:"id"`
 	Titolo      string  `json:"titolo" validate:"required,max=255" example:"Il silenzio di un mare in tempesta"`
 	Autore      string  `json:"autore" validate:"required,max=64" example:"Ruben Fabrizi"`
-	Prezzo      float32 `json:"prezzo" validate:"required" example:"15.90"`
+	Prezzo      float64 `json:"prezzo" validate:"required" example:"15.90"`
 	Summary     string  `json:"summary" validate:"required,max=512" example:"Lorem Ipsum is simply dummy text of the printing and typesetting industry."`
 	Copertina   string  `json:"copertina" validate:"required,eq=Hard Cover|eq=Flexible Cover" example:"Hard Cover"`
 	Genere      string  `json:"genere" validate:"required,eq=Action|eq=Adventure|eq=Business|eq=Cookbooks|eq=Drama|eq=Detective|eq=Fantasy|eq=Fiction|eq=History|eq=Horror|eq=Romance|eq=Psychology|eq=Science Fiction|eq=Short Stories|eq=Thriller" example:"Romance"`
